Check integer values before conversion
- Fix gosec integer overflow conversion warnings (G115/CWE-190)
- Use the `int` type for primary keys in `datatype` structs to avoid
  having to convert values returned from the database layer
- Check that user supplied `int` values are positive before converting
  them to `uint` values
- Validate that the `a3m.Config.AipCompressionLevel` value is between
  zero and nine [0-9]

[skip-codecov]

+++ internal/am/job_tracker_test.go
@@ -25,7 +25,7 @@ import (
 func TestJobTracker(t *testing.T) {
 	t.Parallel()
 
-	paID := uint(1)
+	paID := 1
 	unitID := uuid.New().String()
 
 	clock := clockwork.NewFakeClock()
